+++ doc/t1/t1.clj
@@ -1,9 +1,8 @@
 (ns t1.t1
   (:require [clj-rest-client.core :refer [defrest]]
-            [clj-http.client :as client]
-            [clojure.spec.alpha :as s]))
+            [clj-http.client :as client]))
 
 (defrest
   {"https://api.github.com"
-   {"organizations" (list-organizations [since (s/nilable pos-int?)] {:basic-auth ["user" "personal-token"]})}})
+   {"organizations" (list-organizations [since [:maybe pos-int?]] {:basic-auth ["user" "personal-token"]})}})