+++ doc/t4/t4.clj
@@ -1,17 +1,15 @@
 (ns t4.t4
   (:require [clj-rest-client.core :refer [defrest]]
-            [clj-rest-client.conform :refer :all]
-            [clj-http.client :as client]
-            [clojure.spec.alpha :as s]))
+            [clj-http.client :as client]))
 
 (defrest
   {["users/{username}/orgs" string?] (list-user-organizations [])
-   "organizations" (list-organizations [since (s/nilable pos-int?)])
+   "organizations" (list-organizations [since [:maybe pos-int?]])
    ["repos/{owner}/{repo}" string? string?]
    {"assignees" (list-assignees [])
     "issues" {["{issue_no}" pos-int?] {GET (get-issue [])
-                                       "assignees" {POST (add-issue-assignees [^:key assignees (s/coll-of string?)])
-                                                    DELETE (remove-issue-assignees [^:key assignees (s/coll-of string?)])}}}}})
+                                       "assignees" {POST (add-issue-assignees [^:key assignees [:sequential string?]])
+                                                    DELETE (remove-issue-assignees [^:key assignees [:sequential string?]])}}}}})
 
 (defn git-client [url user pass]
   (fn [req]