+++ doc/t4/t4_end.clj
@@ -1,27 +1,31 @@
 (ns t4.t4-end
   (:require [clj-rest-client.core :refer [defrest]]
-            [clj-rest-client.conform :refer :all]
             [clj-http.client :as client]
-            [clojure.spec.alpha :as s])
-  (:import (java.time.format DateTimeFormatter)))
+            [malli.transform :as mt]))
 
-(s/def ::since (->date-format (DateTimeFormatter/ISO_INSTANT)))
-(s/def ::all-since #(or (not= (:state %) "all") (:since %)))
+(def xf
+  (mt/transformer
+    {:name :rest-client
+     :encoders (select-keys mt/+string-encoders+ ['inst?])
+     :decoders (select-keys mt/+string-decoders+ ['inst?])}))
 
+(def filtered-function [:fn {:error/message "filters that target state \"all\" must limit by \"since\""}
+                        '(fn [{:keys [since state]}] (or (not= state "all") since))])
 
 (defrest
   {["users/{username}/orgs" string?] (list-user-organizations [])
-   "organizations" (list-organizations [since (s/nilable pos-int?)])
+   "organizations" (list-organizations [since [:maybe pos-int?]])
    ["repos/{owner}/{repo}" string? string?]
    {"assignees" (list-assignees [])
-    "issues" {GET (list-repo-issues ::all-since [state (s/nilable string?) since (s/nilable ::since)])
+    "issues" {GET (list-repo-issues filtered-function [state [:maybe string?] since [:maybe inst?]])
               ["{issue_no}" pos-int?] {GET (get-issue [])
-                                       "assignees" {POST (add-issue-assignees [^:key assignees (s/coll-of string?)])
-                                                    DELETE (remove-issue-assignees [^:key assignees (s/coll-of string?)])}}}}})
+                                       "assignees" {POST (add-issue-assignees [^:key assignees [:sequential string?]])
+                                                    DELETE (remove-issue-assignees [^:key assignees [:sequential string?]])}}}}}
+  :transformer xf)
 
 (defn git-client [url user pass]
   (fn [req]
     (-> req
-      (assoc :basic-auth [user pass])
-      (update :url #(str url "/" %))
-      client/request)))
+        (assoc :basic-auth [user pass])
+        (update :url #(str url "/" %))
+        client/request)))