+++ doc/t6/t6_reorg.clj
@@ -1,6 +1,6 @@
 (ns t6.t6-reorg
   (:require [clj-rest-client.core :refer [defrest]]
-            [clj-rest-client.conform :refer :all]
+            [clj-rest-client.transform :refer :all]
             [clj-http.client :as client]
             [clojure.spec.alpha :as s])
   (:import (java.time.format DateTimeFormatter)))