+++ src/clj_rest_client/edn.clj
@@ -1,38 +1,28 @@
 (ns clj-rest-client.edn
-  (:require [clojure.java.io :as io]
-            [clojure.string :refer [starts-with?]]
-            [clojure.edn :as edn]
-            [clojure.walk :as walk])
+  (:require [aero.core :as aero]
+            [clojure.java.io :as io]
+            [clojure.string :refer [starts-with?]])
   (:import (java.net URL)))
 
-(defrecord Ref [key])
-
-(defn ref-obj [key]
-  (->Ref key))
-
-(defn kw-ns [kw]
+(defmethod aero/reader 'crc/ns
+  [_ _ value]
   (let [aliases (assoc (ns-aliases *ns*) (symbol "") *ns*)
-        str-ns (str (namespace kw))
+        str-ns (str (namespace value))
         real-ns (str (get aliases (symbol str-ns) str-ns))]
-    (keyword real-ns (name kw))))
-
-(def default-readers
-  {'crc/ns kw-ns
-   'crc/ref ref-obj})
-
-(defn resolve-refs [all ds]
-  (walk/postwalk #(if (instance? Ref %) (resolve-refs all ((:key %) all)) %) ds))
+    (keyword real-ns (name value))))
 
 (defn load-from-url
-  "Load EDN denoted by the name parameter, using the map of extra readers, and then
-  retrieve the key edn-key
+  "Load EDN denoted by the name parameter using Aero. It will load
+
+  Uses the map of aero opts, and then
+  retrieves the key edn-key or returns whole if edn-key is not defined.
 
   Add -Djava.protocol.handler.pkgs=org.my.protocols to enable custom protocols"
-  [name readers edn-key]
+  [name aero-opts edn-key]
   (when name
-    (let [edn
-          (edn/read-string {:readers (merge default-readers readers)}
-            (slurp (if (starts-with? name "classpath:") (io/resource (subs name 10)) (URL. name))))
-          resolved (resolve-refs edn edn)]
+    (let [edn (aero/read-config (if (starts-with? name "classpath:")
+                                  (io/resource (subs name 10))
+                                  (URL. name))
+                                (or aero-opts {}))]
 
-      (if edn-key (edn-key resolved) resolved))))
+      (if edn-key (edn-key edn) edn))))